requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
folium==0.19.4
geopandas==1.0.1
graphviz==0.20.3
matplotlib==3.10.0
networkx==3.2.1
numpy==2.2.1
pandas==2.2.3
pgmpy==0.1.26
plotly==5.24.1
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.15.1
sentence_transformers==3.3.1
Shapely==2.0.6
streamlit==1.41.1
torch==2.5.1+cu118
datasets
peft
soilgrids